Fixing Merge Conflicts.
