--- src/main/java/controllers/EditHabitController.java
@@ -1,22 +1,0 @@
-package controllers;
-
-import entities.User;
-import use_cases.habits.edit_habit.EditHabitInputBoundary;
-
-public class EditHabitController {
-
-    /**
-     * Create a new habit. Return whether the habit was successfully created.
-     * @param u - The user for the habits
-     * @param s - The name of the habit
-     * @param t - The type of the habit
-     * @param b - The reminder state of the habit
-     */
-    public static void editHabitC(User u, String s, String a, String t, String b){
-
-        boolean r = b.equals("ON");
-
-        EditHabitInputBoundary.editHabitInputBoundary(u, s, a, t, r);
-
-    }
-}
