--- src/main/java/controllers/HabitCreationFailed.java
@@ -1,17 +1,0 @@
-package controllers.habits;
-
-/**
- * Exception deals with failure of creating habit.
- */
-public class HabitCreationFailed extends RuntimeException{
-
-    /**
-     *
-     * @param error - Message to be displayed when creation of habit failed.
-     */
-  public HabitCreationFailed(String error){
-            super(error);
-  }
-
-}
-
