+++ src/main/java/example_user/UserExample.java
@@ -12,19 +12,19 @@ public class UserExample {
 
     static Timer t = new Timer();
 
-    static User person = new User("hello123@gmail.com", "12345678", "Harry", "Potter",
-            hFactory, tFactory, cFactory, t);
+    //static User person = new User("hello123@gmail.com", "12345678", "Harry", "Potter",
+       //     hFactory, tFactory, cFactory, t);
 
-    public static User sendUser(){
-        return person;
-    }
+    //public static User sendUser(){
+        //return person;
+    //}
 
-    public static void send(){
-        System.out.println(person.getHabitCollection().habitCollection.toString());
-    }
+    //public static void send(){
+        //System.out.println(person.getHabitCollection().habitCollection.toString());
+    //}
 
-    public static void main(String[] args) {
-        javax.swing.SwingUtilities.invokeLater(UserExample::send);
-    }
+    //public static void main(String[] args) {
+        //javax.swing.SwingUtilities.invokeLater(UserExample::send);
+    //}
 
 }
